requirements.txt
asgiref==3.6.0
autopep8==2.0.1
dirt-tongue==0.1.2
Django==3.2.16
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2022.7.1
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
